Move the middleware function to a separate file
